% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_create-model-prediction.R
\name{create_model_prediction}
\alias{create_model_prediction}
\title{Create prediction model}
\usage{
create_model_prediction(target_test_df, model_fit)
}
\arguments{
\item{target_test_df}{A dataframe with columns relevant to the regression model, i.e., target variable and predictors}

\item{model_fit}{A "workflow" class object generated by the \code{create_fit} function}
}
\value{
A dataframe with predicted values appended to target_test_df
}
\description{
Create prediction model
}
\examples{
train_df <- target_df(mtcars[1:16, ], 'gear', c("am", "vs"))
test_df <- target_df(mtcars[17:32, ], 'gear', c("am", "vs"))
x_recipe <- create_recipe(train_df, target_variable="gear")
x_spec_list <- create_spec_kmin(train_df, model_recipe=x_recipe, method="lm", target_variable="gear")
x_spec <- get_list_item(x_spec_list, n=1)
x_fit <- create_fit(x_recipe, x_spec, train_df)
create_model_prediction(test_df, x_fit )

}
