% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_create-recipe.R
\name{create_recipe}
\alias{create_recipe}
\title{Create a recipe object for modeling}
\usage{
create_recipe(target_df, target_variable)
}
\arguments{
\item{target_df}{A data frame containing the target variable and predictors.}

\item{target_variable}{A string specifying the name of the target variable.}
}
\value{
A recipe object for modeling.
}
\description{
This function creates a recipe object for modeling based on the specified target variable and predictors.
}
\examples{
data("mtcars")
target_df <- target_df(mtcars, "gear", mpg, cyl, disp)
recipe <- create_recipe(target_df, "gear")

}
