% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_create-spec-kmin.R
\name{create_spec_kmin}
\alias{create_spec_kmin}
\title{Create model specification with optional kmin tuning}
\usage{
create_spec_kmin(
  df,
  model_recipe,
  method,
  kmin = "NA",
  metric,
  target_variable,
  weight_func = "rectangular",
  mode = "regression"
)
}
\arguments{
\item{df}{A data frame containing the training data.}

\item{model_recipe}{A recipe object created using the \code{create_recipe} function.}

\item{method}{A character string indicating the type of regression method to be used: "lm" for linear regression or "kknn" for k-nearest neighbor regression.}

\item{kmin}{A numeric value specifying the minimum number of neighbors to be considered when performing k-nearest neighbor regression. If set to "NA", the function performs a grid search to find the optimal k value. Default is "NA".}

\item{metric}{A character string specifying the performance metric to calculate ("rmse", "rsq", or "mae")}

\item{target_variable}{A character string indicating the name of the target variable to be predicted.}

\item{weight_func}{A character string indicating the weight function used for the k-nearest neighbor regression. Default is "rectangular".}

\item{mode}{A character string indicating the type of regression task. Default is "regression".}
}
\value{
A list containing the model specification and the kmin value (if applicable).
}
\description{
This function creates a model specification for either linear regression or k-nearest neighbor regression, with an optional kmin tuning for the latter. If kmin is not specified, the function performs a grid search to find the optimal k value that gives the minimum root mean squared error (RMSE) on a 5-fold cross-validation of the training data.
}
\examples{
train_df <- mtcars[1:16, ]
target_df <- target_df(train_df, "gear")
model_recipe <- create_recipe(target_df, "gear")
create_spec_kmin(train_df, model_recipe, "lm", metric="rmse", target_variable="gear")
create_spec_kmin(train_df, model_recipe, "kknn", metric="rmse", kmin=5, target_variable="gear")

}
