% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_get-list-item.R
\name{get_list_item}
\alias{get_list_item}
\title{Get List Item}
\usage{
get_list_item(list_object, n)
}
\arguments{
\item{list_object}{A list.}

\item{n}{The index of the element to retrieve.}
}
\value{
The nth element of the input list.
}
\description{
This function retrieves the nth element of a given list.
}
\examples{
my_list <- list(a = 1, b = 2, c = 3)
get_list_item(my_list, 2)

}
