% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_str-collapse.R
\name{str_collapse}
\alias{str_collapse}
\title{Collapse string of arguments and join them by '+'}
\usage{
str_collapse(str_vector)
}
\arguments{
\item{str_vector}{a vector of strings to be collapsed}
}
\value{
A string with the input parameters combined as a string, separated by '+'
}
\description{
Collapse string of arguments and join them by '+'
}
\examples{
str_collapse("my name")
str_collapse(c("my name", "is", "jake"))
str_collapse(c("my name", "is", 2))
str_collapse(c(3, 2, "boy"))

}
