% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_target-df.R
\name{target_df}
\alias{target_df}
\title{Selects target variable and additional columns from a dataframe}
\usage{
target_df(df, target_variable, ...)
}
\arguments{
\item{df}{a dataframe}

\item{target_variable}{a character string representing the target variable}

\item{...}{additional columns to select from the dataframe}
}
\value{
a dataframe with only the target variable and the additional columns selected
}
\description{
Selects target variable and additional columns from a dataframe
}
\examples{
# Selecting only the target variable
target_df(mtcars, "mpg")

# Selecting the target variable and additional columns
target_df(mtcars, "mpg", cyl, disp)

}
